Authentication added successfully

+++ controllers/home_controller.js
@@ -4,21 +4,31 @@ const habbitStatuses = require("../models/habbit_status");
 const ObjectId = mongoose.Types.ObjectId;
 
 module.exports.home= async(req,res)=>{
-    habbits = await Habbit.find();
-    return res.render("index",{
-        title:"Habitesy",
-        habbits:habbits
-})}
+    if(req.user){
+        habbits = await Habbit.find({user:req.user._id});
+        return res.render("home",{
+            title:"Habitesy",
+            habbits:habbits
+    })}else{
+        return res.render("home",{title:"Habitesy"})
+    }
+
+    }
 
 module.exports.create = async(req,res)=>{
     try{
         // adding new habbit to database
-        const habbit = await Habbit.create({
-            title:req.body.name,
-            time:req.body.time,
-            goal:req.body.goal
-        })
-        res.redirect("back");
+        if(req.user){
+            const habbit = await Habbit.create({
+                title:req.body.name,
+                user:req.user._id,
+                time:req.body.time,
+                goal:req.body.goal
+            })
+            req.flash("success","New Habbit added successfully!")
+            res.redirect("back");
+
+        }
     }catch(err){
         console.log("----Error", err);
     }
@@ -27,11 +37,14 @@ module.exports.create = async(req,res)=>{
 module.exports.destroyHabbit = async(req,res)=>{
     try{
         // Deleting the selected habbit from habbit collection
-        await Habbit.deleteOne({_id:req.query.id});
-        // Deleting all the status associated with a habbit
-        await habbitStatuses.deleteMany({habbit:req.query.id});
-        console.log("Habbit deleted successfully");
-        res.redirect("back");
+        if(req.user){
+            await Habbit.deleteOne({_id:req.query.id});
+            // Deleting all the status associated with a habbit
+            await habbitStatuses.deleteMany({habbit:req.query.id});
+            console.log("Habbit deleted successfully");
+            req.flash("success","Habbit deletetion successful")
+            res.redirect("back");
+        }
     }catch(error){
         console.log("Error", err);
     }
@@ -43,14 +56,14 @@ module.exports.toggleStatus = async(req,res)=>{
         const {date, habbit_id, stat} = req.body
         // if the status is the default one then we are not storing that data in the database, hence all those data is deleted for memory optimization
         if(stat==0){
-            await habbitStatuses.deleteOne({habbit:habbit_id, date:date});
+            await habbitStatuses.deleteOne({habbit:habbit_id, date:date, user:req.user._id});
             return res.status(200).json({
                 message:"success",
                 updated:true
             })
         }
         // Creating/ Updating the status value (0,1,2) in the database 
-        let habbitStatus = await habbitStatuses.findOne({habbit:habbit_id, date:date})
+        let habbitStatus = await habbitStatuses.findOne({habbit:habbit_id, date:date,user:req.user._id})
         if (habbitStatus){
             habbitStatus.stat = stat;
             habbitStatus.save();
@@ -59,7 +72,8 @@ module.exports.toggleStatus = async(req,res)=>{
             await habbitStatuses.create({
                 habbit:new ObjectId(habbit_id),
                 date:date,
-                stat:stat
+                stat:stat,
+                user:req.user._id
              })
         }
         return res.status(200).json({
@@ -92,8 +106,15 @@ module.exports.getAllStatus = async(req,res)=>{
 module.exports.getStatCount = async(req,res)=>{
     try{
         const {habbit_id} = req.params;
-        const data = await habbitStatuses.find({habbit:habbit_id, stat:1});
-        return res.status(200).json({success:true, data:data.length})
+        if(req.user){
+            const data = await habbitStatuses.find({habbit:habbit_id, stat:1, user:req.user._id});
+            if(data!=[]){
+                resp=data.length
+            }else{
+                resp = 0
+            }
+            return res.status(200).json({success:true, data:resp})
+        }
     }catch (err){
         console.log("Error in getting completed status count", err)
     }
